refactor: introduce Cgroup interface and consolidate cgroupfs operations
Promotes the cgroup helpers in cgroup/Cgroup.kt to a Cgroup interface
with a CgroupV2 implementation that holds FileSystem in its constructor.
Higher layers (Create, Delete, Ps, runMainProcess) now take a Cgroup
parameter alongside Syscall/FileSystem and call cgroup.setup / cleanup /
getPids instead of free helpers.

This is the same pattern as Syscall (#19/#20) and FileSystem (#22):
domain code goes through an injected interface so tests can substitute
a fake without touching the real cgroupfs.

What changes:
- cgroup/Cgroup.kt: now just the interface (setup, cleanup, getPids).
- cgroup/CgroupV2.kt (new): implementation. Keeps the existing logic
  including the subtree_control ancestor walk (#17 fix), the cgroup v1
  shares -> cgroup v2 weight conversion, the swap.max = swap - limit
  semantic, and the procs write.
- utils/FileSystem.kt gains removeDirectory(path: String): Boolean so
  cgroup cleanup can drop its direct platform.posix.rmdir/access usage.
  RealFileSystem implements it via posix rmdir, FakeFileSystem mutates
  its in-memory directories set.
- Main.kt instantiates CgroupV2(fs) once and threads it into create /
  delete / ps. start / kill / state don't need it.
- Tests:
  - CgroupTest now exercises CgroupV2 directly. New cleanup tests cover
    the no-op-for-null path and FakeFileSystem-backed directory removal.
  - FakeCgroup (new) records calls; available for upcoming higher-layer
    tests of Create / Delete / Ps.

No behavior change is intended; integration workflow exercises mount,
pivot_root, capabilities, seccomp, and cgroup limits end-to-end.
